chore: mv modules around
