+++ src/main.rs
@@ -2,24 +2,17 @@ use std::process::ExitCode;
 
 use anyhow::Result;
 use clap::Parser;
+use ffs::{ffs::scanner::Scanner, types};
 use human_panic::setup_panic;
 use serde::Serialize;
 
-use crate::{
-    ffs::scanner::Scanner,
-    types::{flag::Flag, language::SupportedLanguage},
-};
-
 use futures::future::join_all;
-
-mod ffs;
-mod types;
 
 #[derive(Parser, Clone, Debug)]
 #[command(author, version, about, long_about = None)]
 pub struct Args {
     #[arg(short, long, value_enum)]
-    pub language: SupportedLanguage,
+    pub language: types::language::SupportedLanguage,
     #[arg(short, long, help = "Path to output file (default: STDOUT)")]
     pub output: Option<String>,
     #[arg(short, long, help = "Path to directory to scan (default: .)")]
@@ -50,33 +43,34 @@ async fn main() -> Result<ExitCode> {
 
     let flipt_client = &flipt::api::ApiClient::new(flipt_config)?;
 
-    let flag_results: Vec<Flag> = join_all(found_flags.into_iter().map(|f| async move {
-        let resp = flipt_client
-            .flags()
-            .get(&flipt::api::flag::FlagGetRequest {
-                namespace_key: Some(f.namespace_key.clone()),
-                key: f.flag_key.clone(),
-            })
-            .await;
+    let flag_results: Vec<types::flag::Flag> =
+        join_all(found_flags.into_iter().map(|f| async move {
+            let resp = flipt_client
+                .flags()
+                .get(&flipt::api::flag::FlagGetRequest {
+                    namespace_key: Some(f.namespace_key.clone()),
+                    key: f.flag_key.clone(),
+                })
+                .await;
 
-        match resp {
-            Ok(_) => None,
-            Err(error) => match error.downcast_ref::<flipt::error::Error>() {
-                Some(flipt::error::Error::Upstream(e)) => {
-                    if e.code == NOT_FOUND_ERROR_CODE {
-                        Some(f)
-                    } else {
-                        None
+            match resp {
+                Ok(_) => None,
+                Err(error) => match error.downcast_ref::<flipt::error::Error>() {
+                    Some(flipt::error::Error::Upstream(e)) => {
+                        if e.code == NOT_FOUND_ERROR_CODE {
+                            Some(f)
+                        } else {
+                            None
+                        }
                     }
-                }
-                _ => None,
-            },
-        }
-    }))
-    .await
-    .into_iter()
-    .flatten()
-    .collect();
+                    _ => None,
+                },
+            }
+        }))
+        .await
+        .into_iter()
+        .flatten()
+        .collect();
 
     let mut out_writer: Box<dyn std::io::Write> = match args.output {
         Some(s) => Box::new(std::fs::File::create(s)?),
@@ -114,5 +108,5 @@ struct Results {
 #[serde(rename_all = "camelCase")]
 struct Error {
     message: String,
-    location: crate::types::flag::Location,
+    location: crate::types::location::Location,
 }
