+++ src/types/flag.rs
@@ -1,5 +1,7 @@
 use serde::{Deserialize, Serialize};
 
+use super::location::Location;
+
 #[derive(Debug, Serialize, Clone, Deserialize)]
 #[serde(rename_all = "camelCase")]
 pub struct Flag {
@@ -14,26 +16,6 @@ impl std::fmt::Display for Flag {
             f,
             "namespace_key: {} flag_key: {} [{}]",
             self.namespace_key, self.flag_key, self.location
-        )
-    }
-}
-
-#[derive(Debug, Serialize, Clone, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct Location {
-    pub file: String,
-    pub start_line: usize,
-    pub start_column: usize,
-    pub end_line: usize,
-    pub end_column: usize,
-}
-
-impl std::fmt::Display for Location {
-    fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
-        write!(
-            f,
-            "file: {} start: [line: {}, col: {}] end: [line: {}, col: {}]",
-            self.file, self.start_line, self.start_column, self.end_line, self.end_column
         )
     }
 }
