+++ src/types/mod.rs
@@ -1,3 +1,3 @@
-pub mod args;
 pub mod flag;
 pub mod language;
+pub mod location;
